Reject XLS targets at password I/O and extract OoxmlEncryption
Reject XLS read/write with an IllegalArgumentException pointing to OOXML-only. The read path catches the POI XLS magic detection in EncryptionInfo and re-throws with a guide-oriented message; the write path inspects the workbook's SpreadsheetVersion before any wrap so the failure is early. EncryptedRoundTripTest covers the two guards, an SXSSF round-trip pinning SpreadsheetVersion (not instanceof) as the predicate, and unifies the local helper names.

Move the encryption flow out of SpreadsheetFactory into OoxmlEncryption (decrypt source, wrap target) and EncryptOnCloseOutputStream (encrypt-on-close). The Factory delegates via OoxmlEncryption.decrypt and OoxmlEncryption.wrapTarget; eight POI crypto imports drop out. EncryptionSpec.toEncryptionInfo becomes public for cross-package access with an "internal" javadoc note.
